extractor/requirements.txt
# AI Credit Decisioning Engine — Extractor
# requirements.txt

# ── Core PDF extraction ─────────────────────────
pdfplumber>=0.10.0          # Text-based PDF extraction + table parsing

# ── Scanned PDF / OCR ──────────────────────────
easyocr>=1.7.0              # OCR engine — supports Hindi, no system deps
pdf2image>=1.16.3           # Converts PDF pages to images for OCR
Pillow>=10.0.0              # Image processing for OCR pre-processing
# Legacy OCR (optional, only if OCR_ENGINE=tesseract in .env):
# pytesseract>=0.3.10

# ── Document formats ────────────────────────────
python-docx>=1.1.0          # DOCX extraction
openpyxl>=3.1.2             # XLSX extraction

# ── LLM APIs ───────────────────────────────────
google-genai>=1.0.0          # Gemini API (primary)
anthropic>=0.34.0            # Claude API (fallback)
openai>=1.40.0               # OpenAI GPT-4 (fallback)

# ── Vector Database ────────────────────────────
faiss-cpu>=1.7.4             # FAISS vector store
numpy>=1.24.0                # Array operations for FAISS

# ── Utilities ───────────────────────────────────
python-dotenv>=1.0.0        # Load .env file for API keys
requests>=2.31.0            # HTTP client for HuggingFace API